Module dw::dwfl

DWFL types.

Structs

AddrInfo
Information about a symbol.

Callbacks
Callbacks used to configure the behavior of a Dwfl.

Dwfl
The base type used when interacting with libdwfl.

DwflRef
A borrowed reference to a Dwfl.

Error
A error returned by DWFL APIs.

FindDebuginfo
The callback responsible for locating the debuginfo of a process.

FindElf
The callback responsible for locating the ELF images of a process.

FrameRef
A reference to a stack frame.

ModuleRef
A reference to a module.

Report
A type used to register segments and modules with a DWFL session.

ThreadRef
A reference to a thread.